eapp updates [Shinomori]
git-svn-id: http://eathena.googlecode.com/svn/branches/eapp@2405 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/map/guild.h
@@ -26,7 +26,7 @@ struct guild_castle *guild_mapname2gc(const char *mapname);
 struct map_session_data *guild_getavailablesd(struct guild &g);
 int guild_getindex(struct guild &g,unsigned long account_id,unsigned long char_id);
 int guild_getposition(struct map_session_data &sd,struct guild &g);
-int guild_payexp(struct map_session_data &sd,int exp);
+int guild_payexp(struct map_session_data &sd,unsigned long exp);
 int guild_getexp(struct map_session_data &sd,int exp); // [Celest]
 
 int guild_create(struct map_session_data &sd,const char *name);
